making sure merged changes compile

+++ oracle/src/target_contract_handler.rs
@@ -8,7 +8,7 @@ use data_request::Outcome;
 #[ext_contract]
 pub trait TargetContractExtern {
     fn set_outcome(request_id: U64, requestor: AccountId, outcome: Outcome, tags: Option<Vec<String>>, final_arbitrator_triggered: bool);
-    fn claim_fee(request_id: U64, fee_percentage: u16);
+    fn claim_fee(request_id: U64, fee_percentage: u32);
 }
 
 #[derive(BorshDeserialize, BorshSerialize, Deserialize, Serialize)]
@@ -20,7 +20,7 @@ impl TargetContract {
     pub fn claim_fee(
         &self,
         request_id: U64,
-        fee_percentage: u16
+        fee_percentage: u32
     ) -> Promise {
         target_contract_extern::claim_fee(
             request_id,
